backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
fastapi==0.121.1
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
PySocks==1.7.1
python-dotenv==1.2.1
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
websocket-client==1.9.0
wsproto==1.3.2
